test/multipath: don't wait for bdevperf to finish
After multipath tests are finished we were waiting up to 40s for
bdevperf to finish.  It doesn't test anything, besided the fact that we
can do IOs for 40s after changing to an optimized path.  To reduce test
time, simply kill bdevperf immediately after multipath tests are done.

There's no need to check the status of the perform_tests RPC, as we also
check bdevperf's exit status, so we're bound to catch errors anyway.

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I73fee92c8bd65d780659332de7ddb5f68c5f14d1
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/16665
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Michal Berger <michal.berger@intel.com>
Reviewed-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>