uring: Fix the zcopy error msg for if socket is not added into group
In the master branch, always see the following error msg if
ZEROCOPY send is enabled:

"Unexpected cmsg level or type!"

The root cause is that we do not correctly call the zcopy handling,
and this patch fixes this issue.

When socket is not added into the group, there is no uring queue
initialized. so we can still use the recvmsg on ERRQUEUE to address
this.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: Id9b98f1a2eb5bb6957cbfdf72c922daf385c17a6
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/16689
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>