util/pipe: Add spdk_pipe_group
Pool the data buffers backing the pipe. When a pipe reaches empty state,
it puts the data buffer into the group's pool. If a pipe needs a data
buffer, it takes one from the pool. Since the pool is a stack, a small
number of data buffers tend to be re-used very frequently.

Change-Id: Ic660a326ce33f2ba99820708171597370414a750
Signed-off-by: Ben Walker <benjamin.walker@intel.com>
Signed-off-by: Jacek Kalwas <jacek.kalwas@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/16992
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Mellanox Build Bot