dpdk_mem_info.py: handle spaces in mempool names
DPDK is OK with spaces in mempool names.  SPDK has
never done it, but latest OCF code uses a name with
a space that gets used as a mempool name. The
dpdk_mem_info.py script was parsing mempool
information assuming there were no spaces in the
name.  So this patch fixes it.

Note: I created a dummy memzone with a space in
its name, and confirmed that works correctly.  So
only mempool names needed to be fixed.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I76d4e6e94852dd697b4be187e6ee625db867867a
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/17316
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Ben Walker <benjamin.walker@intel.com>
Reviewed-by: Amir Haroush <amir.haroush@huawei.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>